+++ src/common.h
@@ -14,9 +14,9 @@
 //  Nolqu Runtime Version
 // ─────────────────────────────────────────────
 #define NQ_VERSION_MAJOR 0
-#define NQ_VERSION_MINOR 7
+#define NQ_VERSION_MINOR 8
 #define NQ_VERSION_PATCH 0
-#define NQ_VERSION       "0.7.0"
+#define NQ_VERSION       "0.8.0-beta"
 #define NQ_LANG_NAME     "Nolqu"
 
 // ─────────────────────────────────────────────